Migrated Gulp related server tasks to its own module and cleaned main gulpfile.
